Only set `%cr0.TS` on kernel entry if extra debug checks are enabled
It isn't necessary to disable the FPU while in kernel mode because SVA
can prevent the kernel from using floating-point instructions. By not
redundantly disabling the FPU, we save a control register write on
kernel entry and avoid causing the user process to fault unnecessarily
the next time it tries to use a floating-point instruction.