Save/restore GS Shadow in sva_runvm()
We neglected to handle this register earlier because some poor design
choices on Intel's part (at least from the programmer's perspective,
although they don't have much of an excuse from a hardware designer's
standpoint either since AMD thought better of this) to make every aspect
of segment register state saved/loaded through the VMCS, *except* for
the GS Shadow register. The Intel manual makes no effort to point out
this important fact while describing the VMCS mechanisms, leaving it
entirely in the programmer's hands to figure out that this needs to be
paid attention to.

This commit adds support in SVA for saving/restoring GS Shadow to SVA
internal memory along with a VM's other non-VMCS-resident registers
(such as the GPRs and several other oddball bits). It also adds support
to the sva_get/setvmreg() intrinsics for accessing this register.

This permits us to port Xen's context-switching of GS Shadow to use SVA
intrinsics intead of direct hardware access, the functionality of which
has actually been broken for some time since we changed Xen's GSS access
helper functions to not directly access the hardware during the earlier
port of PV functionality.