Distinguish between error conditions in sva_read/writevmcs()
Xen cares about the difference between the VMfailValid and VMfailInvalid
cases of VMREAD/VMWRITE failure, so we need to distinguish accordingly.
Previously we just returned a generic error and relied on console
printouts to figure out what was going on, which was good enough for the
toy hypervisor from our last paper but not for Xen.