Properly convert 32-bit VMCS bit structs to 64-bit uints for read/write
In one of the more subtly mysterious issues we've had to debug yet, this
was causing VM entry to fail due to the pin-based execution controls
VMCS field being set to all ones (0xffffffff). Further investigation
showed that the compiler was, somehow, concluding that these VMCS reads
would invariably return constant values (0xffffffff in this field's
case, but other incorrect constants like 0x0 for the others), and
thereafter propagating those constants to the subsequent writes back to
those fields.

The root cause turned out to be that we were incorrectly
reinterpret-casting pointers to these bitfield structs, which are 4
bytes wide due to the underlying VMCS fields being defined in the Intel
manual as 32-bit, as pointers to 64-bit uints when passing them to
read/writevmcs_unchecked(). This was easy to miss because the
VMREAD/VMWRITE instructions always deal with 64-bit %rXX registers when
running in long mode, even though the underlying VMCS fields vary in
width (the processor drops the excess upper bits on writes, and clears
them to 0 on reads). Since incorrect casting like this is undefined
behavior in C, the compiler was allowed to make nonsensical assumptions
about the values in question when promoting these address-taken memory
references to register variables across code inlined from
read/writevmcs_unchecked().

Replacing the fast-and-loose type punning with explicit union
constructions fixed the problem by correctly expressing to the compiler
that we do, in fact, need to treat these values both as 32-bit bitfield
structs for access to the individual bits, and as 64-bit uints when
passing them to VMREAD/VMWRITE.

(Debugging and fixing this issue was joint work by me and Colin aka
@AstralSorcerer in a live pair programming session.)