Add locking to Shade intrinsics
We now have a lock field in the VM descriptor structure to prevent race
conditions between multiple processors calling Shade intrinsics to
access the same VM at the same time.

Per Intel, only one logical processor can have a particular VMCS loaded
at a time, so this is in line with how a hypervisor is already expected
to behave. In most cases, we simply return failure from the intrinsic
(rather than spinning to wait to acquire the lock) since lock
acquisition should only fail if the hypervisor is doing something it
shouldn't. Xen has its own higher-level synchronization mechanisms to
ensure sensible sharing of these data structures between different
processors. Shade's locking is merely to prevent Xen from violating our
security guarantees in the event it fails to synchronize itself
correctly (e.g., if an attacker compromises Xen and attempts to create a
race condition).

Note: in order to make adding the lock to the VM descriptor simpler, we
changed the struct vm_desc_t interface slightly such that it is no
longer cleared to all zeroes with a memset() each time a descriptor is
deallocated in sva_freevm(). This was excessive and unnecessarily
expensive; I originally wrote it that way as a "sanity check", but it
ended up encouraging some slightly sloppy code practices since
sva_allocvm() could implicitly assume all descriptor fields were
initialized to zero without explicitly initializing them. This became
evident when, after removing the memset(), the vlAPIC code stopped
working because it expected the vlAPIC mode enum field to be OFF (0) for
a new VM, resulting in a double free of the vlAPIC frames of the
previous VM to hold the descriptor slot when the new VM first tried to
activate its APIC. We now explicitly initialize the vlAPIC mode field in
sva_allocvm(), as we do for other VM descriptor fields and should've
been doing for all of them to keep things clear and up-front.