Add locking to EPT Shade intrinsics
These were missed in the last commit due to being in a different source
file.

Note that since the locking code must check whether the specified VMID
is currently active on the CPU, it was necessary to convert the
host_state structure from a static (file-local) definition to a global
so that it could be accessed from both vmx.c and vmx_ept.c. This is
temporary as we will be shortly moving host_state into the per-CPU
struct CPUState to support simultaneous use on multiple processors.