Re-enable interrupts when performing OS callbacks for secure memory
It turns out Xen doesn't like it very much when you ask it to allocate a
page from its heap while interrupts are disabled. Xen partitions its
locks into "IRQ-safe" and "IRQ-unsafe" categories to prevent deadlock
during IRQ rendezvous. Xen's heap allocation code, which it uses to
provide memory to SVA's callback, needs to take one of the "IRQ-unsafe"
locks. It therefore bugchecks (in debug builds) when it detects it is
attempting to take such a lock with interrupts disabled.

We didn't notice this bug until now because (under Xen) the secure
memory allocator is only used by Shade/VMX code, and our test cases
thusfar haven't allocated enough VMs simultaneously to drain SVA's frame
cache and force it to be refilled during execution of a Shade intrinsic
with interrupts disabled. The frame cache was being initially filled
when Xen called sva_initvmx() on the first CPU during boot, and we were
neglecting to disable interrupts during sva_initvmx() (a separate
issue, a fix to which will be committed separately). Hence, we weren't
encountering this problem in practice...until I added that missing
interrupt-disable code to sva_initvmx().

(FreeBSD 9 didn't seem to surface any problems with this, but that
could've just been because we were running it in single-processor mode
and thus locking would've been much simpler.)

We now re-enable interrupts (if they are disabled) when calling back
into the OS to request or return memory. This is OK from a security
perspective so long as SVA code is careful to not leave the system in an
inconsistent/unsafe state when calling alloc_frame() or free_frame().

(The comments documenting alloc_frame() and free_frame()'s interfaces
have been amended to include this new consideration, and I've addressed
these concerns at all of their call sites within SVA.)