Make the active VM pointer a per-CPU field and save host state on stack
The host_state structure was our last remaining non-thread-safe global
variable in Shade. Most of it, it turns out, can just as well go on the
stack as a local allocation within run_vm() (although we still group it
together as a struct to simplify the entry/exit assembly block). The
only remaining member of host_state that needs to stay as a per-CPU
field is the active VM pointer, which now resides in struct CPUState.