Remove 'volatile' from %gs-relative assembly access in getCPUState()
'volatile' on an asm block prevents the compiler from deleting it during
optimizations. However, as this assembly has no side effects, there is
no reason to have such a restriction. We want the compiler to be able to
optimize accesses to per-CPU data by remembering the result returned by
getCPUState() in a register if appropriate.

Ideally, we would use the __seg_gs pointer annotation to instruct the
compiler to use the GS segment for an ordinary C pointer dereference,
instead of inline assembly. However, when we tried that, the compiler
complained of a null pointer dereference, since the CPUState pointer is
the first field in SVA's TLS area and thus the constant TLS_CPUSTATE
evaluates to 0. There *should* be a way to inform the compiler that we
do in fact intend to dereference %gs:0 here, but until we can figure
that out, a non-volatile asm block works just as well.