Make sure "fields initialized" flag is cleared when a VM slot is re-used
Now that we are no longer memset()'ing the entire VM descriptor each
time a VM is freed (see commit 9529e9a), this flag doesn't get cleared
automatically when a VM descriptor slot gets re-used by a new vCPU. We
now explicitly initialize it to 0 when creating a new VM.

We should have been doing this all along as a matter of software
engineering good-form, since a "dead" VM descriptor slot is logically
uninitialized data after being released by sva_freevm() - and therefore
we should not be counting on its fields being set any particular way
when we re-use that slot in sva_allocvm(). (This is on the principle
that it's better to keep assumptions as contained as possible, rather
than relying on potentially complicated reasoning about how a piece of
freed memory might have been used in the past.)

So that we don't get burned by any more bugs resulting from this, I
reviewed all the struct vm_desc_t fields to find any others that were
relying on assumptions of behavior on previous uses of the descriptor.
The only one that jumped out was the "is_launched" flag, which I've
fixed per the comment in the code.

I also moved our call to xinit() for initializing a new VM's XSAVE area
into sva_allocvm() since it logically fits better there than in
sva_loadvm(). This was, in fact, the bit of code that brought to our
attention that the "fields initialized" flag wasn't being properly
cleared on re-use of a descriptor slot. (At its previous location in
sva_loadvm(), the xinit() call was contingent on the initialized flag
being set properly; without that, we could #GP due to trying to XRSTOR
bad data onto the processor before VM entry.)