Use SVA direct map pointers to page table entries
We were passing the pointer the kernel gave us to functions which expect
a pointer to the PTE being modified. Since the kernel could change the
translation of this pointer at any time, this could potentially allow
the kernel to trick SVA into making an improper page table update. Now,
we translate the kernel pointer to a physical address and make any
accesses through SVA's direct map. Since the kernel pointer is only
translated once, this prevents a potential race with another update
confusing SVA about which page table entry is being modified.