load_eptable_internal(): Fix bug in refcount decrement for old PTP
This bug has been around since the Shade 1.0 days, but "by accident" it
didn't break anything. We were decrementing the refcount on the *new*
PTP instead of the old one, since the code occurred after the point
where we installed the new PTP into vm_descs[vmid].eptp.

The reason this "worked by accident" and didn't cause a crash is that
since we were still only decrementing the refcount once over the course
of a PTP's lifespan (albeit too early in that lifespan), the refcount
math remained consistent overall. Nonetheless, this would have been a
security bug, as it would have allowed the kernel/hypervisor to create
other mappings to the PTP and thus to make changes to it without going
through our page table update intrinsics.