Restore getProcessorID() functionality
This function has been present in SVA but unused (and broken) for a long
time. Re-implemented (in far less convoluted, and more performant,
fashion) in preparation for enhancing VMCS locking to keep track of
*which* processor owns the lock instead of just that *some* processor
owns it. (That'll be necessary to add support for lazy VMCS unloading,
which Xen requires.)

Note that this function still doesn't have any callers yet so it hasn't
exactly been tested...but it should be safe to commit since it doesn't
regress anything. It will be soon to put to use in a subsequent commit
implementing the plans described above.