Use processor ID instead of 1/0 for VM descriptor lock value
This allows us to keep track of *which* processor holds a VM's lock.
This is important in order to support lazy VMCS unloading, since we need
to determine whether it's safe to allow a processor to call sva_loadvm()
or sva_unloadvm() for a VM that is not currently active but whose lock
is nonetheless held. (That's coming soon in another commit.)