Add support for lazy VMCS unloading
Xen takes advantage of a subtlety of the Intel manual that wasn't
immediately obvious to me when I first designed the Shade virtual
instruction set: it's possible to VMPTRLD a new VMCS *without* having to
first unload one that's currently loaded. There can still only ever be
one *active* VMCS at a time (i.e., upon which instructions like
VMLAUNCH/VMRESUME and VMREAD/VMWRITE will act), but multiple VMCSes can
remain *loaded* in the processor's cache without having to be fully
flushed to memory with VMCLEAR.

It turns out that this is extremely important for performance. When we
disabled lazy VMCS unloading in Xen (i.e., had it VMCLEAR the VMCS on
every context switch), performance slowed down by more than 4x. Ouch!
:-< This makes sense in retrospect, as the VMCS is a large structure and
gains a lot from being cached "close to the action". Hypervisor context
switches occur often enough to benefit greatly from keeping all the
VMCSes in cache that are in the context switching working set.

In order to support lazy VMCS unloading, we enhance the VM descriptor
lock to track *which* processor "owns" the VM instead of just that
*some* processor owns it. That allows us to determine whether it is safe
to re-load (i.e., make active again) a VM that is still locked because
it wasn't unloaded when a different VMCS was loaded.