Add support for debug logging of VMCS accesses
To aid in porting Xen's VMCS accesses and implementing runtime checks
for them, it's helpful to be able to profile which fields are actually
used by Xen in practice (and what sorts of values it writes to them).
This commit adds a clumsy, but serviceable mechanism for doing so.

Note that the essential bits to activate the new code are committed with
`#if 0` guards around them; you'll need to flip those to `#if 1` to
enable these debugging features. The reason for this is that the logging
code was not implemented with any regard for performance and is on the
critical path for VM entry/exit handling.