Add intrinsics for VPID and EPT TLB flushes
We already had internal functions for performing all-contexts flushes;
those have been generalized into full-blown intrinsics as needed for the
Xen port.

In the course of doing so, I discovered that there has been a
semi-harmless bug in our internal all-context VPID/EPT flush code since
I first added it a while back. Although it's not especially obvious from
the Intel manual, the processor lays out the 128-bit INVVPID and INVEPT
descriptors as singular little-endian 128-bit integers in memory (which
are in turn broken down bitwise into sub-fields) - not as packed structs
containing adjacent (individually little-endian) integer fields in
normal reading order. This meant that we were reversing the order of the
fields. In the particular case of an all-context flush, this was
harmless, because all the fields are set to 0 anyway. But the code was
nonetheless wrong, and this led to problems when it was generalized to
other types of INVVPID/INVEPT operations where the fields' values
actually matter.

NOTE: the intrinsic sva_flush_vpid_addr() is currently present, but
unimplemented (it will panic if executed), because even though Xen has
code that calls it, I haven't yet encountered a situation where that
code actually executes. It will be straightforward to implement but I
haven't needed to get around to it yet (and perhaps won't, since it's
possible Xen only uses it to support features and modes that we don't
care about).