Disable hack intrinsics for converting between VMIDs and VMCS paddrs
These super ugly intrinsics were a workaround for earlier stages of the
Xen port where Xen needed to retain access to the actual physical
address of the VMCS to perform some operations directly. Now all those
operations have been ported to use Shade intrinsics and we can finally
be rid of these ugly hacks. (As we've done with similar hack itnrinsics
in the past, I've nonetheless kept them in the code, in disabled form,
for reference and in case they are helpful in the future.)