Enable VMCS read checks in log-only mode
Added a bunch of fields to the read whitelist that Xen requires (or
for which it was obvious they should be on the list even if we haven't
logged Xen using them so far).

We *should* be all set on read checks now. Nothing came up in the logs
as having failed a read check after booting, running a unit test, and
running a Linux guest.

I'm leaving these in log-only mode for now, but it would probably be
safe to switch them to hard-fail. I'll probably do so once we've closed
the remaining outstanding issues with the write checks.

Note: this commit changes the VMCS read/write logging facility to log
only accesses that failed security checks by default. Previously, they
logged everything by default; the setting could (and still can) be
changed using the companion userspace command-line utility via a control
hypercall from the PV dom0. This change allows us to ensure that we
aren't failing checks on any VMCS accesses prior to the login screen
coming up.