Initialize VMCS controls to safe defaults instead of provided ones
In the original version of Shade as designed for the VEE '19 paper, I
thought it would be a good idea to require callers of sva_allocvm() to
provide initial values for VMCS controls and guest state fields
up-front, to ensure that the guest can never be started in an undefined
state.

However, this turned out to be unnecessarily prescriptive, and a bad fit
for Xen's vCPU-creation workflow, which figures out what it's going to
put in these fields at a much later time than when it first allocates
memory for the VMCS (i.e., when it calls sva_allocvm()). Shoehorning Xen
into the original sva_allocvm() interface would have had it simply
providing a bunch of dummy values which would get overwritten later by
sva_writevmcs() or sva_setvmreg() calls as appropriate.

The vast majority of these fields don't actually have to be set in any
explicit way to ensure SVA's security guarantees are upheld, as we allow
most of them to be fully under the control of the hypervisor (i.e.,
sva_writevmcs() allows writes to them unconditionally without checking
the values being written). The only remaining concern would be about
data leaking from previously-freed VMs into new ones via uninitialized
memory contents, but that would only be an issue for ghosting VMs;
if/when we add support for those in the future, we'll probably fix that
by just memset()'ing the whole vm_desc_t to 0 in sva_allocvm() (as we
used to do in Shade 1.0, but subsequently removed as it was
unnecessary). (In fact, we still do this for VMCS frames, although -
again - we probably don't *need* to do so for non-ghosting VMs.)

The only VMCS initialization we *need* to ensure to uphold our security
guarantees are the small handful of control fields for which we enforce
specific bitwise settings in sva_writevmcs(). We now simply write safe
defaults to those fields the first time the VMCS is loaded, the same as
we were already doing for other fields which SVA does not allow the
hypervisor to control.

Likewise for non-VMCS state, the only fields that need to be explicitly
initialized to prevent problems are a few header fields in the XSAVE
area, which could cause a #GP if incorrectly set when we try to load it
onto the processor before running the guest. (We still don't do anything
to prevent the hypervisor from explicitly setting them with
sva_setvmfpu() so as to cause a #GP, but that's OK since we do not
attempt to prohibit the system software from hard-crashing the system.
The only reason we bother to prevent this in the initial setting is
because - unlike with the scalar registers - Xen's toolstack never
actually explicitly initializes the FPU state, relying instead on the
fact that Xen's - and now SVA's - initialization of that region will set
it to a sensible "fresh booted" state for a nascent vCPU.)

A positive side effect of this design change is that it allows us to
eliminate/simplify a whole raft of ugly code related to VM
initialization. Most of this code had been disused since the start of
the Xen port since we had to disable it for lack of Xen providing
meaningful initial vaules. With this commit, we have now formalized and
closed the safety gaps on what had been our de facto approach all along,
which was to expect Xen to overwrite meaningless initial values with
meaningful ones before running the guest for the first time.