Fix use-after-free in `sva_reinit_stack`
Previously, we allowed the compiler to place the target in memory, which
it would allocate on the stack. However, the function's stack frame
effectively dies when we reset the stack pointer, which occurs before we
actually jump to the target, resulting in a use-after-free. Errornous
behavior due to this is observed if we take an interrupt in the window
where the stack is dead but before the jump: the memory where the
compiler was placing the target address can get overwritten by the
interrupt handler. We fix this by ensuring the target is placed into a
register.