Add some debug intrinsics to help with frame declaration issues
These proved useful while porting Xen's extended paging code to Shade.
I'm committing them since they may be of general use for debugging frame
type/refcount issues, which are a common category of bugs in SVA and
associated ports.

A live call to sva_check_frame_type() has been left in the Xen port (in
a commit I'm about to make) because it's useful as an assertion to help
track down the root causes of issues where Xen fails to undeclare a PTP
before returning it to its heap. (Such an issue can manifest quite
unintuitively without a check like this at the time the PTP is freed,
because it won't show up until Xen tries to reuse the heap frame
elsewhere, which can happen in any part of the codebase.) This check
shouldn't cause any security issues, but in principle it could be
removed in a "production"/no-asserts version if anyone cared.