Safely initialize EPT root pointer for newly-created VM guests
In the original Shade design (VEE '19 version), the hypervisor was
expected to provide an initial EPT root pointer at the time of VM
creation to ensure that the field was never left uninitialized. This
turned out to not be a good fit for Xen, as Xen treats SVA VM descriptor
allocation as one and the same with VMCS allocation - which happens much
earlier in vCPU setup than EPT (p2m) setup. It's easier for Xen if it
can create the VMCS first and then set its EPT root pointer later.

This is consistent with other adaptations we've made to the Shade
interface to be more flexible and easier for Xen to work with. The
general pattern is that the hypervisor should be able to create the VM
(VMCS) as a single operation without providing any initial settings; SVA
will provide safe defaults for all settings, and the hypervisor can
thereafter change those as it will using the interfaces SVA already
provides for getting/setting them. This design decision now applies
uniformly to guest registers, VMCS fields, and the EPT root pointer, all
of which were explicitly initialized as parameters passed to
sva_allocvm() in the Shade 1.0 design. Now sva_allocvm() takes no
parameters, and the VM creation process is quite a bit "cleaner"
overall.

As a safe default for the EPT root pointer, we create a special
page-table page filled with all zero (i.e. invalid) mappings at boot
within SVA internal memory. All newly-created VMs' EPT root pointers
will point to that page until the hypervisor specifies something
different by calling the sva_load_eptable() intrinsic.

Having made this change, this commit also switches the VMCS read/write
checks to hard-fail mode as intended in the design, since Xen no longer
needs to bypass the sva_read/writevmcs() intrinsics for any VMCS fields.
(The EPT root pointer field was the last such one.)